+++ packages/calypso-url/src/index.ts
@@ -2,4 +2,4 @@ export { URL_TYPE, determineUrlType } from './url-type';
 export { default as format } from './format';
 export { getUrlParts, getUrlFromParts } from './url-parts';
 export { safeImageUrl } from './safe-image-url';
-export { getCaplysoUrl } from './get-calypso-url';
+export { getCalypsoUrl } from './get-calypso-url';