+++ packages/calypso-url/CHANGELOG.md
@@ -4,4 +4,4 @@
 
 ## 1.1.0
 
-- added `getCaplysoUrl` helper
+- added `getCalypsoUrl` helper